requirements.txt
streamlit==1.24.0
streamlit-authenticator==0.3.0
pandas
numpy
requests
jinja2
pyyaml
python-dateutil
weasyprint==55.0
pdfplumber
openpyxl